Add initial code.
Rationale:
=========

This code represents the consolidation of previous work into a Kotlin
Multiplatform Library container such that JVM, JavaScript and macos
platforms are targeted. Android and iOS will be added to the mix soon.

The "feature" for this library is that it is multiplatform Kotlin
code.  Further the JVM target is 100% code covered. JS and macos are
0% code covered, but the fact that most of the code is common means
that in reality, both are highly covered.